+++ R/DataClasses.R
@@ -222,25 +222,9 @@ setClass("Spectrum",
              scanIndex = integer(),
              mz = numeric(),
              intensity = numeric()),
-         validity = function(object) {
-             msg <- validMsg(NULL, NULL)
-             if (any(is.na(intensity(object))))
-                 msg <- validMsg(msg, "'NA' intensities found.")
-             if (any(is.na(mz(object))))
-                 msg <- validMsg(msg, "'NA' M/Z found.")
-             if (any(intensity(object) < 0))
-                 msg <- validMsg(msg, "Negative intensities found.")
-             if (any(mz(object)<0))
-                 msg <- validMsg(msg, "Negative M/Z found.")
-             if (length(object@mz) != length(object@intensity))
-                 msg <- validMsg(msg, "Unequal number of MZ and intensity values.")
-             if (length(mz(object)) != peaksCount(object))
-                 msg <- validMsg(msg, "Peaks count does not match up with number of MZ values.")
-             if (any(diff(mz(object)) < 0))
-                 msg <- validMsg(msg, "MZ values are out of order.")
-             if (is.null(msg)) TRUE
-             else msg
-         })
+         validity = function(object)
+             validSpectrum(object)
+         )
 
 setClass("Spectrum2",
          representation = representation(