+++ NEWS.md
@@ -1,7 +1,8 @@
 # MSnbase 2.1
 
 ## Changes in version 2.1.9
-- (nothing for now)
+- When fully trimmed, an (empty) spectrum has peaksCount of 0L - see
+  https://github.com/lgatto/MSnbase/issues/184 <2017-01-20 Fri>
 
 ## Changes in version 2.1.8
 - Remove order option altogether <2017-01-19 Thu> (superseeds setting